+++ src/constants.ts
@@ -1,1 +1,0 @@
-export const FILE_URI_PREFIX_RE = /^file:\/\//;